+++ src/lib/visa/visa.engineer.spec.ts
@@ -1,4 +1,7 @@
-import { EngineerQualifications, calculatePoints } from '@lib/visa/b'
+import {
+  EngineerQualifications,
+  calculatePoints,
+} from '@lib/visa/visa.engineer'
 import 'jest-extended'
 import { errorMessages } from './errors'
